## Summary
Extracts `ConnectionPackageLoader` from `ConnectionQuery` as part of the cleanup initiative (#3756).

Closes #3777.

### Changes

- **Created `ConnectionPackageLoader.cs`** — focused class containing:
  - `LoadPackagesByKeyAsync` — loads assignment, role, and delegation packages
  - `EnrichPackageResourcesAsync` — attaches resources to loaded packages

- **Created `ConnectionIndex.cs`** — extracted the generic `ConnectionIndex<T>` helper into its own file

- **Simplified `ConnectionQuery.cs`**:
  - Delegates to `_packageLoader` fields
  - Public API unchanged

### Constraints met

- ✅ No change in packages
- ✅ `ConnectionQuery` public API unchanged
- ✅ All 305 integration tests pass
