+++ src/apps/Altinn.AccessManagement/src/Altinn.AccessMgmt.PersistenceEF/Queries/Connection/ConnectionIndex.cs
@@ -1,0 +1,33 @@
+﻿using Altinn.AccessMgmt.PersistenceEF.Queries.Connection.Models;
+
+namespace Altinn.AccessMgmt.PersistenceEF.Queries.Connection;
+
+internal sealed class ConnectionIndex<T>
+{
+    private readonly Dictionary<ConnectionCompositeKey, List<T>> map = new();
+
+    public void Add(ConnectionCompositeKey key, T item)
+    {
+        if (!map.TryGetValue(key, out var list))
+        {
+            map[key] = list = new List<T>(4);
+        }
+
+        list.Add(item);
+    }
+
+    public void AddRange(ConnectionCompositeKey key, IEnumerable<T> items)
+    {
+        if (!map.TryGetValue(key, out var list))
+        {
+            map[key] = list = new List<T>();
+        }
+
+        list.AddRange(items);
+    }
+
+    public IReadOnlyList<T> Get(ConnectionCompositeKey key) =>
+        map.TryGetValue(key, out var list) ? list : Array.Empty<T>();
+
+    public IEnumerable<KeyValuePair<ConnectionCompositeKey, List<T>>> Pairs => map;
+}
